Move email tests to SynchronizerIntegrationTests
* Move email user state tests to SynchronizerIntegrationTests
